requirements.txt
openai-whisper
soundfile
numpy
librosa
nvgpu